modify

+++ src/main/java/com/hy/lang/mercury/pojo/enums/OrderStatus.java
@@ -1,23 +1,23 @@
 package com.hy.lang.mercury.pojo.enums;
 
 public enum OrderStatus {
-    已下单(10),
-    已支付(20),
-    已发货(30),
-    已收货(40),
-    已取消(90);
+    已下单(10L),
+    已支付(20L),
+    已发货(30L),
+    已收货(40L),
+    已取消(90L);
 
-    private int code;
+    private Long code;
 
-    OrderStatus(int code) {
+    OrderStatus(Long code) {
         this.code = code;
     }
 
-    public int getCode() {
+    public Long getCode() {
         return code;
     }
 
-    public void setCode(int code) {
+    public void setCode(Long code) {
         this.code = code;
     }
 }
